removed falsehood that is irrelevant to list
While it's cool and all to acknowledge that some people are not male or female, and that some people are not comfortable with the way they were born; I feel that this Item has no place in a list about programmer falsehoods.

Perhaps if the list were a a curated list of falsehoods in general, not specific to the programmer niche it would fit, otherwise; what stops other miscellaneous falsehoods from making it to the list? Perhaps someone could submit a PR later including articles explaining that adding salt to water does not make it boil quicker, it would be just as relevant.